deleted unused files and cleaned used ones
