+++ modules/high_level_functions_CO1.py
@@ -1,18 +1,19 @@
 import ee 
 
-#from . import configs
-from modules import utils_string
-from modules import ms_indices_C02 as indices
-from modules import utils_Landsat_SR_C02 as utils_LS
+
+from . import configs
+from . import utils_string
+from . import ms_indices_CO2 as indices
+from . import utils_Landsat_SR_CO1 as utils_LS
 
 
 def makeLandsatSeriesSrFiltered(config):
     # 1. load Landsat data and calculate indices
     collection = utils_LS.makeLandsatSeriesSr(config['geom'], config['date_filter_yr'], config['date_filter_mth'], config['meta_filter_cld'])\
-    .map(indices.ndvi) \
-    .map(indices.ndmi) \
-    .map(indices.ndwi) \
-    .map(indices.tc) 
+    .map(indices.ndvi57) \
+    .map(indices.ndmi57) \
+    .map(indices.ndwi57) \
+    .map(indices.tc5)
 
     # 2. Filter pixels off 3 std from mean
     std_diff = utils_LS.calculate_std_diff(collection, 3)
@@ -31,17 +32,16 @@ def add_external_mask(external_mask):
         return image.updateMask(combined_mask)
     return wrap
 
-
 def runTCTrend(config_trend):
   # 1. load Landsat data and calculate indices
   collection = utils_LS.makeLandsatSeriesSr(config_trend['geom'], 
                                             config_trend['date_filter_yr'], 
                                             config_trend['date_filter_mth'], 
                                             config_trend['meta_filter_cld']) \
-  .map(indices.ndvi) \
-  .map(indices.ndmi) \
-  .map(indices.ndwi) \
-  .map(indices.tc)
+  .map(indices.ndvi57) \
+  .map(indices.ndmi57) \
+  .map(indices.ndwi57) \
+  .map(indices.tc5)
 
   if 'mask' in config_trend.keys():
       if isinstance(config_trend['mask'], ee.Image):
@@ -57,39 +57,40 @@ def runTCTrend(config_trend):
   print(config_trend['select_bands_visible'])
   def mask_outliers(image):
       return utils_LS.update_mask_by_std(image, lower, upper, config_trend['select_bands_visible'])
-  #collection = collection.map(mask_outliers)
-  
-  # TODO: calculate annaul mosaic here
-  # TODO: This part here breaks the Collection
+  collection = collection.map(mask_outliers)
 
   # 3. Calculate image pixel count
   image_observations = collection.count().select([1], ['nObservations'])
   #image_total_count = collection.count().select([0], ['imageCount'])
 
-  # 4. Calculate trend 
+  # 4. Calculate trend
   trend_image = ee.Image()
   print(config_trend['select_indices'])
   for index in config_trend['select_indices']:
-    # TODO: change 'Date' to 'Year'
     trend = ee.ImageCollection(collection.select(['Date', index])) \
       .reduce(ee.Reducer.linearFit().unweighted()) \
       .select(['scale', 'offset', 'scale', 'scale'], 
               [index + '_slope', index + '_offset', index + '_upper', index + '_lower'])
-    trend_image = trend_image.addBands(trend).clip(config_trend['geom'])
+    trend_image = trend_image.addBands(trend)
 
-  # TODO: change factor to 10
   trend_image = trend_image.multiply(ee.Image.constant(3650))
 
-  # 6. Create visual output 
+  # 5. Calculate basic collection statistics
+  #
+  
+  # 6. Create visual output
+  #trend_image = trend_image
   trend_image_visual = trend_image.select(config_trend['select_TCtrend_bands']) \
-                                  .unitScale(-0.12, 0.12)\
+                                  .unitScale(-1200, 1200)\
                                   .multiply(ee.Image.constant(255)).uint8() # Scale to values from -0.12 to 0.12 \
      
+
   return {'visual': trend_image_visual,
           'data': trend_image,
           'image_collection': collection,
           'n_observations': image_observations.uint16()
   }
+
 
 
 def exportTCTrendImage(config):
